third_party/blink/web_tests/svg/custom/rect-calc-corner-radii-expected.txt

CONSOLE ERROR: line 4: Error: <rect> attribute ry: A negative value is not valid. ("-100")
